requirements.txt
python-telegram-bot==21.4
requests
pandas
numpy
rapidfuzz
pydantic

python-dotenv
